Added some missing includes and added some typecasts for 64-bit mode. Icons from ttf now checks if file exists before adding. Tweaked premake files to allow Engine and SharedGameCode files to exists in different folder configurations.

+++ MyEngine/SourceCommon/ComponentSystem/EngineComponents/ComponentMonoScript.cpp
@@ -367,7 +367,7 @@ void ComponentMonoScript::AddAllVariablesToWatchPanel()
                             }
 
                             if( strcmp( className, m_MonoClassName ) == 0 )
-                                currentClassIndex = validClasses.size();
+                                currentClassIndex = (int)validClasses.size();
 
                             validClasses.push_back( className );
                         }
